requirements.txt
langchain-openai
python-dotenv
langchain-community
pypdf
langchain-core
streamlit